+++ management-api-agent-dse-6.9/src/main/java/com/datastax/mgmtapi/shim/DseAPI69.java
@@ -32,6 +32,7 @@ import java.util.concurrent.Callable;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import org.apache.cassandra.auth.IRoleManager;
+import org.apache.cassandra.auth.RoleResource;
 import org.apache.cassandra.concurrent.TPCTaskType;
 import org.apache.cassandra.config.DatabaseDescriptor;
 import org.apache.cassandra.cql3.QueryProcessor;
@@ -337,5 +338,24 @@ public class DseAPI69 implements CassandraAPI {
 
   public void reloadInternodeEncryptionTruststore() throws Exception {
     DseReloadableTrustManager.serverEncryptionInstance().reloadTrustManager();
+  }
+
+  @Override
+  public List<Map<String, String>> listRoles() {
+    IRoleManager roleManager = getRoleManager();
+    Set<RoleResource> allRoles = roleManager.getAllRoles();
+    List<Map<String, String>> roles = new ArrayList<>();
+    for (RoleResource role : allRoles) {
+      Map<String, String> roleOutput = new HashMap<>();
+      roleOutput.put("name", role.getRoleName());
+      roleOutput.put("super", String.valueOf(roleManager.isSuper(role)));
+      roleOutput.put("login", String.valueOf(roleManager.canLogin(role)));
+
+      roleOutput.put("options", "{}");
+      roleOutput.put("datacenters", "");
+      roles.add(roleOutput);
+    }
+
+    return roles;
   }
 }