+++ management-api-server/src/main/java/com/datastax/mgmtapi/resources/AuthResources.java
@@ -5,30 +5,43 @@
  */
 package com.datastax.mgmtapi.resources;
 
+import static com.datastax.mgmtapi.resources.v1.TableOpsResources.LIST_OF_MAP_OF_STRINGS;
+
 import com.datastax.mgmtapi.ManagementApplication;
 import com.datastax.mgmtapi.resources.common.BaseResources;
+import com.datastax.mgmtapi.resources.models.Role;
+import com.datastax.oss.driver.api.core.cql.ResultSet;
+import com.datastax.oss.driver.api.core.cql.Row;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.QueryParam;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 import org.apache.commons.lang3.StringUtils;
 
 @Path("/api/v0/ops/auth")
 public class AuthResources extends BaseResources {
+  private static final ObjectMapper jsonMapper = new ObjectMapper();
 
   public AuthResources(ManagementApplication application) {
     super(application);
   }
 
   @POST
-  @Path("/role/")
+  @Path("/role")
   @Operation(summary = "Creates a new user role", operationId = "createRole")
   @Produces(MediaType.TEXT_PLAIN)
   @ApiResponse(
@@ -71,6 +84,86 @@ public class AuthResources extends BaseResources {
               password);
 
           return Response.ok("OK").build();
+        });
+  }
+
+  @DELETE
+  @Path("/role")
+  @Operation(summary = "Drops a user role", operationId = "dropRole")
+  @Produces(MediaType.TEXT_PLAIN)
+  @ApiResponse(
+      responseCode = "200",
+      description = "Role dropped",
+      content =
+          @Content(
+              mediaType = MediaType.TEXT_PLAIN,
+              schema = @Schema(implementation = String.class),
+              examples = @ExampleObject(value = "OK")))
+  @ApiResponse(
+      responseCode = "400",
+      description = "Username is empty",
+      content =
+          @Content(
+              mediaType = MediaType.TEXT_PLAIN,
+              schema = @Schema(implementation = String.class),
+              examples = @ExampleObject(value = "Username is empty")))
+  public Response dropRole(@QueryParam(value = "username") String name) {
+    return handle(
+        () -> {
+          if (StringUtils.isBlank(name))
+            return Response.status(Response.Status.BAD_REQUEST.getStatusCode(), "Username is empty")
+                .build();
+
+          app.cqlService.executePreparedStatement(
+              app.dbUnixSocketFile, "CALL NodeOps.dropRole(?)", name);
+
+          return Response.ok("OK").build();
+        });
+  }
+
+  @GET
+  @Path("/role")
+  @Operation(summary = "Lists existing user rules", operationId = "listRoles")
+  @Produces(MediaType.TEXT_PLAIN)
+  @ApiResponse(
+      responseCode = "200",
+      description = "List of roles",
+      content =
+          @Content(
+              mediaType = MediaType.APPLICATION_JSON,
+              array = @ArraySchema(schema = @Schema(implementation = Role.class))))
+  @ApiResponse(
+      responseCode = "400",
+      description = "Username is empty",
+      content =
+          @Content(
+              mediaType = MediaType.TEXT_PLAIN,
+              schema = @Schema(implementation = String.class),
+              examples = @ExampleObject(value = "Username is empty")))
+  public Response listRoles() {
+    return handle(
+        () -> {
+          List<Role> roleResult = new ArrayList<>();
+
+          ResultSet result =
+              app.cqlService.executePreparedStatement(
+                  app.dbUnixSocketFile, "CALL NodeOps.listRoles()");
+          Row row = result.one();
+          if (row != null) {
+            List<Map<String, String>> roles = row.get(0, LIST_OF_MAP_OF_STRINGS);
+
+            for (Map<String, String> role : roles) {
+              roleResult.add(
+                  new Role(
+                      role.get("name"),
+                      Boolean.valueOf(role.get("super")),
+                      Boolean.valueOf(role.get("login")),
+                      role.get("datacenters"),
+                      role.get("options")));
+            }
+          }
+
+          return Response.ok(roleResult, MediaType.APPLICATION_JSON).build();
         });
   }
 }